Share code between scripts
- Refactor scripts structure
