+++ tools/internals/models/game_properties.py
